Refactor geometric_kernels.kernels.feature_maps into the subpackage geometric_kernels.feature_maps, move geometric_kernels.sampling.probability_densities there.
